# backend/app/schemas.py
from datetime import datetime
from uuid import UUID

from pydantic import BaseModel, ConfigDict

from machineplay.schemas import GameStatus, GameStreamEvent


class StartGameRequest(BaseModel):
    white_engine_id: UUID
    black_engine_id: UUID
    runner_id: UUID


class StartGameResponse(BaseModel):
    id: UUID
    status: str
    white: UUID
    black: UUID


class RunnerOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    runner_id: UUID
    name: str


class UserOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: UUID
    github_id: int
    login: str
    name: str | None
    avatar_url: str
    is_admin: bool
    created_at: datetime


class EngineOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: UUID
    name: str
    command: str
    description: str


class LiveStreamEvent(BaseModel):
    game_id: UUID
    event: GameStreamEvent


class GameOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: UUID
    white_id: UUID
    black_id: UUID
    white_name: str
    black_name: str
    status: GameStatus
    result: str | None
    moves: list[str]
    fen: str
    pgn: str | None
    white_clock: float
    black_clock: float
    created_at: datetime
    ended_at: datetime | None